requirements.txt
fancy-config~=0.12.0
rin-jsonutils~=1.1.0
redis~=4.3.4